Memoria 64

+++ src/components/memory.cpp
@@ -11,7 +11,7 @@ uint64_t Memory::align_addr(uint64_t address) const {
     return (address / BLOCK_BYTES) * BLOCK_BYTES;
 }
 
-void Memory::read_block(uint64_t address, uint8_t* out_block) const {
+void Memory::read_block(uint64_t address, uint64_t* out_block) const {
     uint64_t base = align_addr(address);
     if (base + BLOCK_BYTES > MEM_BYTES) {
         throw std::out_of_range("Memory::read_block: address out of range");
@@ -21,7 +21,7 @@ void Memory::read_block(uint64_t address, uint8_t* out_block) const {
     std::cout << "[MEM] read_block @ 0x" << std::hex << base << std::dec << "\n";
 }
 
-void Memory::write_block(uint64_t address, const uint8_t* in_block) {
+void Memory::write_block(uint64_t address, const uint64_t* in_block) {
     uint64_t base = align_addr(address);
     if (base + BLOCK_BYTES > MEM_BYTES) {
         throw std::out_of_range("Memory::write_block: address out of range");
@@ -30,7 +30,7 @@ void Memory::write_block(uint64_t address, const uint8_t* in_block) {
     std::cout << "[MEM] write_block @ 0x" << std::hex << base << std::dec << "\n";
 }
 
-void Memory::read_bytes(uint64_t address, uint8_t* out_buf, size_t n) const {
+void Memory::read_bytes(uint64_t address, uint64_t* out_buf, size_t n) const {
     if (address + n > MEM_BYTES) throw std::out_of_range("Memory::read_bytes out of range");
     std::memcpy(out_buf, mem_.data() + address, n);
 }
